% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbind.R
\name{fbind}
\alias{fbind}
\title{Bind factors nicely}
\usage{
fbind(a, b)
}
\arguments{
\item{a}{A factor}

\item{b}{A factor}
}
\value{
A factor
}
\description{
This makes nice factor combos
}
\examples{
a <- factor(c("hi", "there"))
b <- factor(c("what's", "up"))
fbind(a, b)
}
