% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbind.R
\name{fbind}
\alias{fbind}
\title{Bind two factors}
\usage{
fbind(a, b)
}
\arguments{
\item{a}{A factor.}

\item{b}{A factor.}
}
\value{
A factor.
}
\description{
I could say more. This is called the 'description'.
}
\examples{
a <- factor(c("character", "hits", "your", "eyeballs"))
b <- factor(c("but", "integer", "where it", "counts"))
fbind(a, b)
}
